Dodano polje tenanta u flat, napravljeno da se samo stanovi bez tenanta mogu brisati

+++ Software/app/src/main/java/hr/foi/rampu/stanarko/database/FlatsDAO.kt
@@ -37,8 +37,9 @@ class FlatsDAO {
                 documents.forEach{
                     val helpVariable = it.toObject(Flat::class.java)
                     if(helpVariable != null){
-
-                        db.collection("flats").document(it.id).delete()
+                        if(helpVariable.tenants.isEmpty()){
+                            db.collection("flats").document(it.id).delete()
+                        }
                     }
                 }
             }
